requirements.txt
diff-match-patch==20181111
lxml==4.3.3
pandas==0.24.2
python-telegram-bot==11.1.0
